requirements.txt
##############################
#### PROJECT REQUIREMENTS ####
##############################
RPi.GPIO
subprocess.run